profile_enhancer/app/requirements.txt
fastapi
uvicorn
requests
bs4
numpy
pandas
cachetools
google-generativeai
